# Conflicts:
#	src/main/kotlin/com/b05/newsfeedproject/config/SwaggerConfig.kt
#	src/main/kotlin/com/b05/newsfeedproject/domain/exception/ModelNotFoundException.kt
#	src/main/kotlin/com/b05/newsfeedproject/domain/user/controller/UserController.kt
#	src/main/kotlin/com/b05/newsfeedproject/domain/user/dto/UserResponse.kt
#	src/main/kotlin/com/b05/newsfeedproject/domain/user/model/User.kt
#	src/main/kotlin/com/b05/newsfeedproject/domain/user/model/admin/AdminInitializer.kt
#	src/main/kotlin/com/b05/newsfeedproject/domain/user/repository/UserRepository.kt
#	src/main/kotlin/com/b05/newsfeedproject/domain/user/service/UserService.kt
#	src/main/kotlin/com/b05/newsfeedproject/security/JwtAuthenticationFilter.kt

+++ src/main/kotlin/com/b05/newsfeedproject/domain/user/model/User.kt
@@ -16,33 +16,33 @@ import java.util.Date
 @Entity
 @Table(name = "app_user")
 class User(
-    @Column(name = "email", nullable = false)
-    var email: String,
-    @Column(name = "password", nullable = false)
-    var password: String,
-    @Column(name = "nickname", nullable = false)
-    var nickName: String,
+        @Column(name = "email", nullable = false)
+        var email:String,
+        @Column(name = "password", nullable = false)
+        var password:String,
+        @Column(name = "nickname", nullable = false)
+        var nickName:String,
 
-    @Enumerated(EnumType.STRING)
-    @Column(name = "type", nullable = false)
-    var type: UserType = UserType.USER,
+        @Enumerated(EnumType.STRING)
+        @Column(name = "type", nullable = false)
+        var type:UserType = UserType.USER,
 
-    @CreationTimestamp
-    @Column(name = "create_at", nullable = true)
-    var create_at: LocalDateTime? = null,
-    @UpdateTimestamp
-    @Column(name = "update_at", nullable = true)
-    var update_at: LocalDateTime? = null
-) {
+        @CreationTimestamp
+        @Column(name = "create_at", nullable = true)
+        var create_at:LocalDateTime? = null,
+        @UpdateTimestamp
+        @Column(name = "update_at", nullable = true)
+        var update_at:LocalDateTime? = null
+){
 
 
     companion object {
-        fun from(request: SignupUserRequest, encoder: PasswordEncoder) = User(
-            email = request.email,
-            password = encoder.encode(request.password),
-            nickName = request.nickName,
-            create_at = LocalDateTime.now(),
-            update_at = null
+        fun from(request:SignupUserRequest,encoder:PasswordEncoder) = User(
+                email = request.email,
+                password = encoder.encode(request.password),
+                nickName = request.nickName,
+                create_at = LocalDateTime.now(),
+                update_at = null
         )
 
     }
@@ -50,11 +50,13 @@ class User(
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    var id: Int? = null
+    var id:Long? = null
+
+
 
 
     //프로필 수정 시 패스워드 암호화
-    fun update(request: UpdateUserRequest, encoder: PasswordEncoder) {
+    fun update(request:UpdateUserRequest,encoder: PasswordEncoder) {
         email = request.email
         password = encoder.encode(request.password)
         nickName = request.nickName
@@ -62,11 +64,11 @@ class User(
 }
 
 fun User.toResponse() = UserResponse(
-    id,
-    email,
-    nickName,
-    password,
-    type.toString(),
-    create_at.toString(),
-    update_at.toString()
+        id,
+        email,
+        nickName,
+        password,
+        type.toString(),
+        create_at.toString(),
+        update_at.toString()
 )
